llama/requirements.txt
flask
flask-cors
llama-cpp-python
huggingface_hub
